==> notebooks/0-1_download_GoogleAudioSet.ipynb <==
# Download the Google AudioSet

import pandas as pd

unbalanced_df = pd.read_csv('../audioset-processing/data/unbalanced_train_segments.csv',header=[0,1])

unbalanced_df

# get the youtube videos labeled as city/urban or rural/nature
urban_ind = unbalanced_df[unbalanced_df.columns[3]].str.contains('/t/dd00128')
natural_ind = unbalanced_df[unbalanced_df.columns[3]].str.contains('/t/dd00129')

df_urban = unbalanced_df[urban_ind]
df_natural = unbalanced_df[natural_ind]

# randomly sampled 10,000 files from more than 20,000 files
df_urban_sampled = df_urban.sample(n=10000, random_state=23, axis=0)
df_urban_sampled

df_natural_sampled = df_natural.sample(n=10000, random_state=23, axis=0)
df_natural_sampled

df = pd.concat([df_urban_sampled, df_natural_sampled])
df

df.to_csv('../audioset-processing/data/unbalanced_train_segments_sampled10000_UrbanNatural.csv', index=False)

# Download files

from distutils.sysconfig import get_python_lib
print(get_python_lib())

import os
os.getcwd()

os.chdir('/Users/andrewchang/Springboard_GitHub/Capstone_AcousticEnv/audioset-processing')
os.getcwd()

Note (Feb 26): 
make this change to the file /Users/andrewchang/opt/anaconda3/envs/AcousticEnv/lib/python3.8/site-packages/youtube_dl/extractor/youtube.py
https://github.com/ytdl-org/youtube-dl/issues/31530
original line: 'uploader_id': self._search_regex(r'/(?:channel|user)/([^/?&#]+)', owner_profile_url, 'uploader id') if owner_profile_url else None,

# nature unbalanced set
! python3 process.py download -c "Outside, rural or natural" -d ../data/raw/GoogleAudioSet_unbalanced_list --label_file data/class_labels_indices.csv --csv_dataset data/unbalanced_train_segments_sampled10000_UrbanNatural.csv

# city unbalanced set
! python3 process.py download -c "Outside, urban or manmade" -d ../data/raw/GoogleAudioSet_unbalanced_list --label_file data/class_labels_indices.csv --csv_dataset data/unbalanced_train_segments_sampled10000_UrbanNatural.csv

# nature evaluation set
! python3 process.py download -c "Outside, rural or natural" -d ../data/raw/GoogleAudioSet_eval_list --label_file data/class_labels_indices.csv --csv_dataset data/eval_segments.csv

# city evaluation set
! python3 process.py download -c "Outside, urban or manmade" -d ../data/raw/GoogleAudioSet_eval_list --label_file data/class_labels_indices.csv --csv_dataset data/eval_segments.csv
