==> notebooks/0-2_download_GoogleAudioSet_excludeMusic.ipynb <==
# Take out the audio files containing music

Based on my observation, files that contain a music label typically involve heavy audio editing, sound effects, or even music that replaces the original audio recording of the environment. As a result, those files should be excluded.

import pandas as pd

def takeout_music(file_dir):
    # file_dir = '../data/raw/GoogleAudioSet_unbalanced_list/Outside, rural or natural'
    import glob
    df = pd.read_csv(file_dir+'.csv', header=None, names=['ID', 'start', 'end', 'label'])
    df['start'] = df['start'].astype('int').astype('str')
    dl = glob.glob(file_dir+'/*')
    dl_id = [i.rsplit('/', 1)[1] for i in dl]
    dl_id = [i.rsplit('.', 2)[0] for i in dl_id]
    df['filename'] = df['ID']+'_'+df['start']
    ind = df['filename'].isin(dl_id)
    df_files = df[ind]
    df_files.drop_duplicates(inplace = True)
    music_ind = df_files['label'].str.contains('/m/04rlf')
    df_noMusic = df_files[~music_ind]
    return df_noMusic

df_nature_noMusic = takeout_music('../data/raw/GoogleAudioSet_unbalanced_list/Outside, rural or natural')
df_city_noMusic = takeout_music('../data/raw/GoogleAudioSet_unbalanced_list/Outside, urban or manmade')
df_nature_eval_noMusic = takeout_music('../data/raw/GoogleAudioSet_eval_list/Outside, rural or natural')
df_city_eval_noMusic = takeout_music('../data/raw/GoogleAudioSet_eval_list/Outside, urban or manmade')

df_nature_noMusic.to_csv('../data/interim/GoogleAudioSet_unbalanced_list/nature_no_music.csv')
df_city_noMusic.to_csv('../data/interim/GoogleAudioSet_unbalanced_list/city_no_music.csv')
df_nature_eval_noMusic.to_csv('../data/interim/GoogleAudioSet_eval_list/nature_no_music.csv')
df_city_eval_noMusic.to_csv('../data/interim/GoogleAudioSet_eval_list/city_no_music.csv')

df_nature_noMusic

df_city_noMusic

df_nature_eval_noMusic

df_city_eval_noMusic
