==> notebooks/0-3_re-evaluate_AudioSet_labels.ipynb <==
# Re-evaluating the labels

Despite the large quantity of files available on Google AudioSet, the quality of its labeling is not great. To get a rough idea of the labeling accuracy of our dataset, I randomly sampled 100 files (50 urban/city and 50 rural/nature) and evaluated whether they were correctly labeled.

import pandas as pd
import numpy as np
rural_pd = pd.read_csv('../data/interim/GoogleAudioSet_unbalanced_list/nature_no_music.csv').sample(n = 50, random_state=23, ignore_index=True)
urban_pd = pd.read_csv('../data/interim/GoogleAudioSet_unbalanced_list/city_no_music.csv').sample(n = 50, random_state=23, ignore_index=True)

rural_pd['urls'] = 'https://www.youtube.com/watch?v='+rural_pd['ID']+'&t='+rural_pd['start'].astype(int).astype(str)
rural_pd['urls']

n = np.nan
rural_pd['label_correct'] = [
    1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 
    1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 
    1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 
    1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 
    1, 1, 1, 1, 1, 1, 1, 1, 1, 1
]
np.nanmean(rural_pd['label_correct'])

urban_pd['urls'] = 'https://www.youtube.com/watch?v='+urban_pd['ID']+'&t='+urban_pd['start'].astype(int).astype(str)
urban_pd['urls']

n = np.nan
urban_pd['label_correct'] = [
    1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 
    1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 
    1, 1, 1, 1, 1, 0, 0, 1, 1, 0,
    0, 1, 1, 0, 1, 1, 1, 0, 0, 0, 
    1, 1, 1, 1, 0, 0, 1, 0, 1, 0
]
    
np.nanmean(urban_pd['label_correct'])

**It turns out that the labeling accuracy was roughly between 64-84%. Based on this, I assume that the overall maximum classification accuracy of the current dataset is 80%.**
